requirements.txt
anthropic
openai
pydantic
python-dotenv
aiohttp
tiktoken